Ok, trying to rational numbers only and avoid angles is not working.
The motivation for using rationals was to have precise representations of vectors that was immune to the problems of comparing floating point numbers. However I'm going through gymnastics to compare these rational numbers like vectors. And the values we are dealing with are on a less than 100x100 cartesian grid where all coordinates are integers. Rounding is not a big deal.

Also this solution was posted to Reddit and it is so much cleaner:

https://github.com/prscoelho/aoc2019/blob/master/src/aoc10/mod.rs

Also the direction of my current solution currently does not preserve the original locations of asteroids, so I need to rework everything to do part 2.